Update 28.2
- CRUD of Orders
- Change Entities
- Change Readme

+++ src/main/java/fpt/sep490/entity/Product.java
@@ -41,7 +41,7 @@ public class Product implements Serializable {
     private Boolean isEnabled;
 
     @Column(name = "cost", nullable = false)
-    private BigInteger cost;
+    private Long cost;
 
     @Column(name = "discount_percent")
     private float discountPercent;
